add in IMU monitor

+++ src/Monitors/TempMonitor.cpp
@@ -30,9 +30,9 @@ void TempMonitor::execute()
     }
 
     // Convert the data
-    uint32_t rawtmp = (data[0] * 256 + data[1]) & 0xFFFC;
-    double temp_c = -46.85 + (175.72 * (rawtmp / 65536.0));
-    double temp_f = temp_c * 1.8 + 32;
+    float rawtmp = (data[0] * 256 + data[1]) & 0xFFFC;
+    float temp_c = -46.85 + (175.72 * (rawtmp / 65536.0));
+    float temp_f = temp_c * 1.8 + 32;
 
     sfr::temperature::temp_c->set_value(temp_c);
     sfr::temperature::temp_f->set_value(temp_f);
